+++ tagger-gui-scala/src/main/scala/Domain.scala
@@ -4,7 +4,7 @@ package no.lau.tagger.scala.model
  * The domain model works as the bridge between the Java and Scala world
  */
 
-class ScalaSegment(val id: String, var start: Float, var end: Float, var text: String)
+class ScalaSegment(val id: String, var start: Int, var end: Int, var text: String)
 
 class ScalaSong(val reference: String, val mediaFile: ScalaMediaFile, var segments: List[ScalaSegment], var bpm: Float) {
   def segmentWithId(id:String):Option[ScalaSegment] = segments.filter(segment=> segment.id == id).headOption
