+++ tagger-gui-scala/src/main/scala/SegmentPlayer.scala
@@ -8,7 +8,7 @@ import no.lau.tagger.scala.model.ScalaSong
 import actors.Actor
 import org.codehaus.httpcache4j.cache.VdvilCacheStuff
 
-class SegmentPlayer(var song: ScalaSong, startCue: Float, endCue: Float) extends Actor {
+class SegmentPlayer(var song: ScalaSong, startCue: Int, endCue: Int) extends Actor {
   val mp3File = VdvilCacheStuff.fileLocation(song.mediaFile.fileName)
   val audioSource: AudioSource = new MP3Source(mp3File)
   val player = new AudioPlayer
@@ -17,8 +17,8 @@ class SegmentPlayer(var song: ScalaSong, startCue: Float, endCue: Float) extends
   def act {
     log.debug("startCue playing = " + startCue)
     var instructions = new Instructions
-    var playLength: Float = endCue - startCue
-    instructions.append(new SimpleAudioInstruction(0F, playLength, song.bpm.floatValue, startCue, song.mediaFile.startingOffset.floatValue, audioSource, 1F))
+    var playLength: Int = endCue - startCue
+    instructions.append(new SimpleAudioInstruction(0, playLength, song.bpm.floatValue, startCue, song.mediaFile.startingOffset.floatValue, audioSource, 1F))
     player.playMusic(instructions)
   }
