+++ tagger-model/src/main/java/no/lau/tagger/model/Segment.java
@@ -5,10 +5,10 @@ import java.io.Serializable;
 public class Segment implements Serializable {
     public final String id;
     public final String text;
-    public final Float start;
-    public final Float end;
+    public final int start;
+    public final int end;
 
-    public Segment(String id, Float start, Float end, String text) {
+    public Segment(String id, int start, int end, String text) {
         this.id = id;
         this.start = start;
         this.end = end;
